requirements.txt
requests
py-socrata
